pyproject.toml
[tool.poetry]
name = "weni-mcp"
version = "0.1.0"
description = "weni mcp"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mcp = "^1.26.0"
httpx = "^0.28.1"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
